Implement Image_Count_Manager class
The current implementation for retrieving view counts creates a new session
every time the function is called, making it become slow. This is especially
so since it needs to be frequently called.

Let's turn it into a class to retain session information.

Manual creation and management of event loops are done to ensure the
event loop is persistent, and the session is always called within
this same event loop to prevent aiohttp errors.